+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Latest:
  - Add `data-pieces` and `parity-pieces` flags to `siac renter upload`
  - Add SIA_DATA_DIR environment variable for setting the data directory for
    siad/siac
+ - Add blacklist for skylinks
  - Made build process deterministic. Moved related scripts into `release-scripts`
  
 **Bugs Fixed**